Add fmtlib to opm_add_tests calls with libraries.
The list of libraries is assumed to contain all needed ones.
Hence we need to list it ecplicitly.

For the object libraries we need to put the fmtlib compile
definitions and include path into the the respective properties
(as older CMake versions do not allow using target_link_libraries
on object libraries).