Added --throw-untabulated=Bool parameter to allow preventing aborts.
For a CO2 model UniformTabulated2d would throw an exception during
`BlackoilWellModel::initializeWellState` when a well was suddenly
switched on.

To prevent that there is now a `--throw-untabulated=Bool` parameter
for flow that allows the user to still get this behavior. The default
value is false, which means a change to previous flow versions that
would always throw.