Fix Averaged Quantities for Non-HC Simulations
Commit 6d3da3d2e introduced the notion of a "dynamic", pressure
dependent pore-volume and switched the 'PoreVolume' aggregates to
reference condition evaluation for output to the 'PORV' field in the
.PRT file.  We did however fail to update all existing uses of
'PoreVolume' which introduced an inconsistency.  In particular, for
simulation models without hydrocarbons-e.g., a single-phase water
run-the numerator of the volume weighted average would include
pressure effects through the rock compressibility but the
denominator would not.

Thanks to Edmund Stephens for discovering the issue and providing an
example model for analysing the underlying issue.