Adjust Whitespace in Output Module
Mostly to split long lines and to add a few blank lines for
readability.  While here, also apply 'const' in a few locations and
flip the order of a conditional to (potentially) save a bit of work.

This is in preparation of revising the output destination categories
of numerous 3D result vectors.