Split 3D Vectors into Base and Extended Categories
The 'base' arrays are compatible in both name and interpretation to
those emitted in the 'SOLUTION' section from other simulators
whereas the 'extended' arrays are specific to Flow.  Tag extended
arrays as OPM_EXTENDED instead of AUXILIARY as the latter is
deprecated and will be removed at some point in the future.