Try to detect more exception during parallel simulator setup.
We have experienced some deadlocks lately because of programming
errors. This change will not result in a clean exit for those
but rather an abort by MPI because of non-matching communications.

Also we cannot forsee whether people will add more exceptions in
those parts that were not protected. We better play it safe here.