Apply Regional Multipliers to Initial Transmissibilities
This commit adds a new member function,

    EclTransmissibility<>::applyNncMultreg_()

which applies regional transmissibility multipliers such as those
entered in the MULTREGT keyword to the explicit input NNCs.  We make
the application conditional on a new parameter, default value
'false', and pass 'true' as the argument from 'finishInit()'.

Along with TransMult::getRegionMultiplierNNC(), this implements all
known connection behaviours for inter-region connections.
Multipliers internal to a region in MULTREGT are not yet supported.