Refacto Fluid-In-Place Calculations For Maintainability
This commit splits updateFluidInPlace_() into several smaller helper
functions, each with a narrow purpose.  They're all just called from
the original call site--the body of updateFluidInPlace_()--but this
new version is, in my opinion, easier to reason about and there is
less shared state.

In anticipation of adding support for summary vectors FHPV and RHPV
(field and region levels of hydrocarbon pore-volumes), we also split
the pore-volume updates out to a branch separate from that needed
for average pressure calculations.