Workaround to compile issue with two tests.
Two tests include ebos.hh, but probably should not: test_glift1 and test_RestartSerialization.
This change restores the build for those tests. Removing ebos.hh from their inclusion list
should be done, but currently exposes that some other headers are not properly standalone,
so it is a bit more work to do.