Treat exceptions during computation of well potentials as warnings.
Numerical problems during the calculation were reported as errors and
counted as such in the PRT file. Yet, the simulation would carry
on (maybe with a smaller timestep), anyway. At least some users are
parsing the PRT file to extract the number errors to get an indication
whether a run succeeded. For this having
```
Error:
[opm-simulators/opm/simulators/wells/StandardWellPrimaryVariables.cpp:722] Infinite primary variable after update from wellState, well: BLA
...
Error summary:
Warnings 1416
Info 980
Errors 1
Bugs 0
Debug 0
Problems 0
```
on an otherwise successful run seemed unreasonable.

Therefore we now make sure that those exceptions are treated as
warnings when they are logged (Errors 0). We do this by adding a
boolean flag to DeferredLogger::logMessage to indicate that we want
this kind of convertion. Default is false.