[fix] Use return code for error detection (not uninitialized MPI_Status::ERROR
According to MPI standard the ERROR field of MPI_Status might not be initialized
unless for operations that return multiple statuses, see  Section 3.7.5
of the standard. In older OpenMPI versions (<=4.0.x) we were lucky
that ERROR was initialized to 0 always. This is not the case for 4.1.y
at least. See https://github.com/open-mpi/ompi/issues/12049.

Therefore we use the retun code to determine whether there was an
error. Note that the default error handler usually is to abort the
application if errors occur. In that case the error code will always
return success.