added: write out RPTSOL configured FIPs to PRT file on simulation start
add a method in EclWriter to enable this.
this is called the first time a call is made to WriteOutput,
as that happens after initial conditions have been applied which
is required to get the proper output.

this also fixes a long-standing issue where the initial FIP state was
taken after the first time step.